chore(seeding): proposition to better type safe
